refactor: move new routes to /nodes
