+++ desci-server/src/routes/v1/doi.ts
@@ -1,21 +1,10 @@
 import { Router } from 'express';
 
-import {
-  asyncHander,
-  attachDoi,
-  attachDoiSchema,
-  checkMintability,
-  ensureNodeAccess,
-  ensureUser,
-  getDoi,
-  mintDoi,
-  validate,
-} from '../../internal.js';
+import { asyncHander, checkMintability, ensureNodeAccess, ensureUser, getDoi, mintDoi } from '../../internal.js';
 const router = Router();
 
 router.post('/check/:uuid', [ensureUser, ensureNodeAccess], asyncHander(checkMintability));
 router.post('/mint/:uuid', [ensureUser, ensureNodeAccess], asyncHander(mintDoi));
-router.post('/attach', [ensureUser, ensureNodeAccess, validate(attachDoiSchema)], asyncHander(attachDoi));
 router.get('/:identifier', [ensureUser], asyncHander(getDoi));
 
 export default router;
